Create README.md
I am an energetic and knowledgeable recruiter (and nurse) and I am excited to work with clients and candidates to play matchmaker and find a good fit for employers and prospective candidates. I am hiring for behavioral health positions throughout New England. In particular: LICSWs, LCSWs, LMHCs, PMHNPs, mental health workers, RNs, LPNs, APRNs. We do permanent placement only, we are not an agency. We support companies hiring from the bedside to corporate positions like program management, CFOs, and VPs. I am more than a recruiter because I want to continue to make a difference as a healthcare professional myself.